Refactor ReportTable stories to Component Story Format.
